docs: 스킬 관련 문서 내용 분리
