+++ space-d/src/docs/asciidoc/index.adoc
@@ -18,13 +18,4 @@ include::{docdir}/domain/today-quiz.adoc[]
 include::{docdir}/domain/local-image.adoc[]
 include::{docdir}/domain/report.adoc[]
 include::{docdir}/domain/bookmark.adoc[]
-
-[[Skill-API]]
-== 스킬 API
-
-[[Find-Skill]]
-=== 스킬 조회
-
-operation::skill-controller-test/스킬_조회_요청_성공_테스트[snippets='http-request,request-headers,http-response,response-fields']
-
-include::{snippets}/docs-controller-test/exceptions/exception-response-fields-findSkillException.adoc[]
+include::{docdir}/domain/skill.adoc[]
