+++ space-d/src/main/resources/static/docs/common/company.html
@@ -444,50 +444,8 @@ body.book #toc,body.book #preamble,body.book h1.sect0,body.book .sect1>h2{page-b
 <div id="preamble">
 <div class="sectionbody">
 <div id="company" class="paragraph">
-<p>Company 허용 값</p>
+<p>Unresolved directive in company.adoc - include::/Users/labtop/intellij/space-d/space-d/build/generated-snippets/docs-controller-test/enums/enum-response-fields-company.adoc[]</p>
 </div>
-<table class="tableblock frame-all grid-all stretch">
-<colgroup>
-<col style="width: 50%;">
-<col style="width: 50%;">
-</colgroup>
-<thead>
-<tr>
-<th class="tableblock halign-left valign-top">이름</th>
-<th class="tableblock halign-left valign-top">의미</th>
-</tr>
-</thead>
-<tbody>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>BLIND</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">비공개</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>MAJOR</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">대기업</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>SMALL</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">중소기업</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>FOREIGN</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">외국계</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>STARTUP</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">스타트업</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>MIDSIZE</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">중견기업</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>JOB_HUNTER_INTERN</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">취준생/인턴</p></td>
-</tr>
-</tbody>
-</table>
 </div>
 </div>
 </div>
