+++ space-d/src/main/resources/static/docs/common/experience.html
@@ -444,50 +444,8 @@ body.book #toc,body.book #preamble,body.book h1.sect0,body.book .sect1>h2{page-b
 <div id="preamble">
 <div class="sectionbody">
 <div id="experience" class="paragraph">
-<p>Experience 허용 값</p>
+<p>Unresolved directive in experience.adoc - include::/Users/labtop/intellij/space-d/space-d/build/generated-snippets/docs-controller-test/enums/enum-response-fields-experience.adoc[]</p>
 </div>
-<table class="tableblock frame-all grid-all stretch">
-<colgroup>
-<col style="width: 50%;">
-<col style="width: 50%;">
-</colgroup>
-<thead>
-<tr>
-<th class="tableblock halign-left valign-top">이름</th>
-<th class="tableblock halign-left valign-top">의미</th>
-</tr>
-</thead>
-<tbody>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>BLIND</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">비공개</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>BETWEEN_SECOND_THIRD</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">2~3년 차</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>BETWEEN_THIRD_FOURTH</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">3~4년 차</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>BETWEEN_FIRST_SECOND</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">1~2년 차</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>OVER_FIFTH</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">5년 차 이상</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>UNDER_FIRST</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">1년 차 미만</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>BETWEEN_FOURTH_FIFTH</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">4~5년 차</p></td>
-</tr>
-</tbody>
-</table>
 </div>
 </div>
 </div>
