+++ space-d/src/main/resources/static/docs/common/profile-image-name.html
@@ -444,54 +444,8 @@ body.book #toc,body.book #preamble,body.book h1.sect0,body.book .sect1>h2{page-b
 <div id="preamble">
 <div class="sectionbody">
 <div id="profile-image-name" class="paragraph">
-<p>ProfileImageName 허용 값</p>
+<p>Unresolved directive in profile-image-name.adoc - include::/Users/labtop/intellij/space-d/space-d/build/generated-snippets/docs-controller-test/enums/enum-response-fields-profileImageName.adoc[]</p>
 </div>
-<table class="tableblock frame-all grid-all stretch">
-<colgroup>
-<col style="width: 50%;">
-<col style="width: 50%;">
-</colgroup>
-<thead>
-<tr>
-<th class="tableblock halign-left valign-top">이름</th>
-<th class="tableblock halign-left valign-top">의미</th>
-</tr>
-</thead>
-<tbody>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>VENUS</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">금성</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>EARTH</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">지구</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>MERCURY</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">수성</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>NEPTUNE</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">해왕성</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>MARS</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">화성</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>SATURN</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">토성</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>URANUS</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">천왕성</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>JUPITER</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">목성</p></td>
-</tr>
-</tbody>
-</table>
 </div>
 </div>
 </div>
