+++ space-d/src/main/resources/static/docs/common/pronunciation-type.html
@@ -444,34 +444,8 @@ body.book #toc,body.book #preamble,body.book h1.sect0,body.book .sect1>h2{page-b
 <div id="preamble">
 <div class="sectionbody">
 <div id="profile-image-name" class="paragraph">
-<p>PronunciationType 허용 값</p>
+<p>Unresolved directive in pronunciation-type.adoc - include::/Users/labtop/intellij/space-d/space-d/build/generated-snippets/docs-controller-test/enums/enum-response-fields-pronunciationType.adoc[]</p>
 </div>
-<table class="tableblock frame-all grid-all stretch">
-<colgroup>
-<col style="width: 50%;">
-<col style="width: 50%;">
-</colgroup>
-<thead>
-<tr>
-<th class="tableblock halign-left valign-top">이름</th>
-<th class="tableblock halign-left valign-top">의미</th>
-</tr>
-</thead>
-<tbody>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>UK</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">영국 발음</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>US</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">미국 발음</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>KOREAN</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">한글 발음</p></td>
-</tr>
-</tbody>
-</table>
 </div>
 </div>
 </div>
