+++ space-d/src/main/resources/static/docs/common/quiz-category.html
@@ -444,38 +444,8 @@ body.book #toc,body.book #preamble,body.book h1.sect0,body.book .sect1>h2{page-b
 <div id="preamble">
 <div class="sectionbody">
 <div id="profile-image-name" class="paragraph">
-<p>QuizCategory 허용 값</p>
+<p>Unresolved directive in quiz-category.adoc - include::/Users/labtop/intellij/space-d/space-d/build/generated-snippets/docs-controller-test/enums/enum-response-fields-quizCategory.adoc[]</p>
 </div>
-<table class="tableblock frame-all grid-all stretch">
-<colgroup>
-<col style="width: 50%;">
-<col style="width: 50%;">
-</colgroup>
-<thead>
-<tr>
-<th class="tableblock halign-left valign-top">이름</th>
-<th class="tableblock halign-left valign-top">의미</th>
-</tr>
-</thead>
-<tbody>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>TOTAL</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">전체 실무</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>DESIGN</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">디자인</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>BUSINESS</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">비즈니스</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>DEVELOP</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">개발</p></td>
-</tr>
-</tbody>
-</table>
 </div>
 </div>
 </div>
