+++ space-d/src/main/resources/static/docs/common/report-reason.html
@@ -444,42 +444,8 @@ body.book #toc,body.book #preamble,body.book h1.sect0,body.book .sect1>h2{page-b
 <div id="preamble">
 <div class="sectionbody">
 <div id="profile-image-name" class="paragraph">
-<p>ReportReason 허용 값</p>
+<p>Unresolved directive in report-reason.adoc - include::/Users/labtop/intellij/space-d/space-d/build/generated-snippets/docs-controller-test/enums/enum-response-fields-reportReason.adoc[]</p>
 </div>
-<table class="tableblock frame-all grid-all stretch">
-<colgroup>
-<col style="width: 50%;">
-<col style="width: 50%;">
-</colgroup>
-<thead>
-<tr>
-<th class="tableblock halign-left valign-top">이름</th>
-<th class="tableblock halign-left valign-top">의미</th>
-</tr>
-</thead>
-<tbody>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>PERSONAL_INFORMATION_EXPOSURE</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">개인정보 노출 위험</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>PROFANITY</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">욕설, 음란 등 부적절한 내용</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>ETC</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">기타</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>OVER_COMMENT</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">댓글 도배</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>SPAM</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">광고 및 홍보성 내용</p></td>
-</tr>
-</tbody>
-</table>
 </div>
 </div>
 </div>
