+++ space-d/src/main/resources/static/docs/common/today-quiz-status.html
@@ -444,34 +444,8 @@ body.book #toc,body.book #preamble,body.book h1.sect0,body.book .sect1>h2{page-b
 <div id="preamble">
 <div class="sectionbody">
 <div id="profile-image-name" class="paragraph">
-<p>TodayQuizStatus 허용 값</p>
+<p>Unresolved directive in today-quiz-status.adoc - include::/Users/labtop/intellij/space-d/space-d/build/generated-snippets/docs-controller-test/enums/enum-response-fields-todayQUizStatus.adoc[]</p>
 </div>
-<table class="tableblock frame-all grid-all stretch">
-<colgroup>
-<col style="width: 50%;">
-<col style="width: 50%;">
-</colgroup>
-<thead>
-<tr>
-<th class="tableblock halign-left valign-top">이름</th>
-<th class="tableblock halign-left valign-top">의미</th>
-</tr>
-</thead>
-<tbody>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>NOT_LOGGED_IN</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">인증 전</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>NOT_SOLVED</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">풀이 전</p></td>
-</tr>
-<tr>
-<td class="tableblock halign-left valign-top"><p class="tableblock"><code>SOLVED</code></p></td>
-<td class="tableblock halign-left valign-top"><p class="tableblock">풀이 후</p></td>
-</tr>
-</tbody>
-</table>
 </div>
 </div>
 </div>
