+++ space-d/src/test/java/com/dnd/spaced/config/docs/snippet/exceptions/skill/SkillExceptionControllerTest.java
@@ -1,0 +1,38 @@
+package com.dnd.spaced.config.docs.snippet.exceptions.skill;
+
+import static org.springframework.restdocs.payload.PayloadDocumentation.beneathPath;
+import static org.springframework.restdocs.snippet.Attributes.attributes;
+import static org.springframework.restdocs.snippet.Attributes.key;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import com.dnd.spaced.config.docs.snippet.CommonExceptionControllerTest;
+import org.junit.jupiter.api.Test;
+import org.springframework.http.MediaType;
+import org.springframework.test.web.servlet.MvcResult;
+import org.springframework.test.web.servlet.ResultActions;
+
+class SkillExceptionControllerTest extends CommonExceptionControllerTest {
+
+    @Test
+    void exceptions() throws Exception {
+        ResultActions resultActions = mockMvc.perform(
+                get("/test/skills/exceptions").contentType(MediaType.APPLICATION_JSON)
+        );
+
+        MvcResult mvcResult = resultActions.andReturn();
+        SkillExceptionDocs data = findExceptionData(mvcResult, SkillExceptionDocs.class);
+
+        resultActions.andExpect(status().isOk())
+                .andDo(
+                        restDocs.document(
+                                customResponseFields(
+                                        "exception-response",
+                                        beneathPath("data.findSkillException").withSubsectionId("findSkillException"),
+                                        attributes(key("title").value("`GET /skills` 예외 상황")),
+                                        exceptionConvertFieldDescriptor(data.findSkillException())
+                                )
+                        )
+                );
+    }
+}
