+++ space-d/src/test/java/com/dnd/spaced/config/docs/snippet/exceptions/skill/SkillExceptionDocs.java
@@ -1,0 +1,9 @@
+package com.dnd.spaced.config.docs.snippet.exceptions.skill;
+
+import com.dnd.spaced.config.docs.snippet.exceptions.ExceptionContent;
+import java.util.Map;
+import lombok.Builder;
+
+@Builder
+public record SkillExceptionDocs(Map<String, ExceptionContent> findSkillException) {
+}
